added new order page style and fixed confirm order button

+++ js/order.js
@@ -1,13 +1,12 @@
-$(document).ready(function () {
-  const shoppingCart = JSON.parse(localStorage['shoppingCart'])
+$(document).ready(function() {
+  const shoppingCart = JSON.parse(localStorage["shoppingCart"]);
   const inventory = JSON.parse(sessionStorage.getItem("inventory"));
-  let orderList = document.querySelector('.order-list')
+  let orderList = document.querySelector(".order-list");
   let subTotal = 0;
   let itemsCountTotal = 0;
-  orderList.innerHTML = ``
+  orderList.innerHTML = ``;
 
   Object.keys(shoppingCart).forEach(itemID => {
-
     const item = inventory.find(item => item.id === Number(itemID));
     const itemCount = shoppingCart[itemID];
     const itemTotal = item.price.value * itemCount;
@@ -19,8 +18,8 @@ $(document).ready(function () {
         <td class="item-name">${item.title}</td>
         <td class="item-qty">${itemCount}</td>
         <td class="item-price">${item.price.value} kr</td>
-        <td class="item-amount">${itemTotal} kr</td>
-      </tr>`
+        <td class="item-total">${itemTotal} kr</td>
+      </tr>`;
   });
 
   orderList.innerHTML += `
@@ -28,13 +27,33 @@ $(document).ready(function () {
       <td>Totalt:</td>
       <td class="products-amount"></td>
       <td></td>
-      <td>${subTotal} kr</td>
-    </tr>
+      <td class="item-total" >${subTotal} kr</td>
+    </tr >
     </tbody>
-    `
+    `;
 
-  document.querySelector('.products-amount').innerHTML = itemsCountTotal
+  document.querySelector(".products-amount").innerHTML = itemsCountTotal;
 
-  //empty localStorage after print out order
-  localStorage.setItem("shoppingCart", JSON.stringify({}));
-})
+  const confirmButton = document.querySelector(".confirm-order-button");
+  if (itemsCountTotal === 0) {
+    confirmButton.setAttribute("aria-disabled", true);
+    confirmButton.classList.add("disabled");
+    confirmButton.setAttribute("disabled", "");
+  } else {
+    confirmButton.setAttribute("aria-disabled", false);
+    confirmButton.classList.remove("disabled");
+    confirmButton.removeAttribute("disabled");
+  }
+
+  confirmButton.addEventListener("click", onOrderConfirmedClick);
+});
+
+function onOrderConfirmedClick(event) {
+  if (confirm("Vill du bekräfta din order?")) {
+    localStorage.setItem("shoppingCart", JSON.stringify({}));
+    alert("Tack för din order!");
+    location.href = "/index.html";
+  } else {
+    event.preventDefault();
+  }
+}
